app/requirements.txt
Django==3.0.6
gunicorn==19.9.0
psycopg2-binary==2.8.3
djangorestframework==3.11.0